style: move the switch button to the sidebar
