+++ src/style.css
@@ -31,6 +31,7 @@
   --text-order: #ffffff;
   --bg-drawer: rgba(0, 0, 0, 0.5);
   --section-bg: rgb(0, 0, 0, 0.5);
+  --icon-switch: #e5bb90;
 }
 
 :root[class="my-app-dark"] {
